Various improvement to 1-Wire Failures
Before if one of the 1-wire thermometers were to fail to rea a temperature, the entire program would go into a failure mode that would result in system no longer detecting hi pressure (the main point of the system). Now we try & catch per temperature read and then set a default if we cannot read a temperature and still run the main loop wile resetting the 1-wire bus.  Additionally, I set the hydronic system to kick in when there are less zones active the colder it gets so that the geothermal system doesn't have to work too hard, this uses more electricity however it protects the system from freezing.