Code Cleanup and Fix geo_electric mode
I cleaned up the main loop to make it easier to read and at the same time fixed when we turn on the electric backup to be not based on the number of zones (and outside temperature) but based on iff the temperature of the buffer tank has gotten too cold (the geo system alone is not able to keep up with demand). This will result in less flip flopping of electric backup (I think).