Add New Pump Control
the following change removes the oil heating system control since I do not have it anymore and adds a new GPIO output pin to control the geothermal pump. the geothermal well pump will need to be on when I am heating in the winter and cooling in the summer as the new exterior heat pump uses the well as a thermal sink in the summer time thus heating up the well.
What is also included in this commit is a shuffling of the temerature sensors since I re-wired them.